Updated My Previous main.py with error handling
Now the program won't error out for some times and end immediatey. This was very irritating for me when I was trying to improve it as every time I tested it, it would just end as soon as I touched the sensor 90% of the time. Now it won't end immediately but will show the error message and start the program again. Now I can test it in peace on Lenovo Laptop!